# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MEI Fácil API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /revenue:
    get:
      operationId: listRevenue
      tags: [revenue]
      summary: List revenue entries
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueListResponse"
    post:
      operationId: createRevenue
      tags: [revenue]
      summary: Create a revenue entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revenue"
  /revenue/{id}:
    get:
      operationId: getRevenue
      tags: [revenue]
      summary: Get a revenue entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revenue"
    put:
      operationId: updateRevenue
      tags: [revenue]
      summary: Update a revenue entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevenueUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revenue"
    delete:
      operationId: deleteRevenue
      tags: [revenue]
      summary: Delete a revenue entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /revenue/stats:
    get:
      operationId: getRevenueStats
      tags: [revenue]
      summary: Get revenue statistics and projections
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueStats"
  /revenue/monthly:
    get:
      operationId: getMonthlyRevenue
      tags: [revenue]
      summary: Get monthly revenue breakdown for chart
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenue"
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      security:
        - bearerAuth: []
      parameters:
        - name: folder
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document (metadata only - file via multipart)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, name, email, cpf, cnpj]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        cpf:
          type: string
        cnpj:
          type: string
        phone:
          type: ["string", "null"]
        openingDate:
          type: ["string", "null"]
        activity:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        annualLimit:
          type: number
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        openingDate:
          type: ["string", "null"]
        activity:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        annualLimit:
          type: ["number", "null"]
        onboardingCompleted:
          type: ["boolean", "null"]
...
    Revenue:
      type: object
      required: [id, userId, amount, date, category, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        date:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    RevenueInput:
      type: object
      required: [amount, date, category]
      properties:
        amount:
          type: number
        date:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
    RevenueUpdate:
      type: object
      properties:
        amount:
          type: ["number", "null"]
        date:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
    RevenueListResponse:
      type: object
      required: [entries, total]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Revenue"
        total:
          type: number
    RevenueStats:
      type: object
      required: [totalAnnual, totalMonthly, avgMonthly, projectedAnnual, usedPercent, annualLimit, monthsWithData]
      properties:
        totalAnnual:
          type: number
        totalMonthly:
          type: number
        avgMonthly:
          type: number
        projectedAnnual:
          type: number
        usedPercent:
          type: number
        annualLimit:
          type: number
        monthsWithData:
          type: integer
        alertLevel:
          type: string
          enum: [safe, warning, danger, exceeded]
    MonthlyRevenue:
      type: object
      required: [month, year, total]
      properties:
        month:
          type: integer
        year:
          type: integer
        total:
          type: number
        label:
          type: string
...
    Document:
      type: object
      required: [id, userId, name, folder, fileType, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        folder:
          type: string
          enum: [das, notas_fiscais, declaracoes, certidoes, outros]
        fileType:
          type: string
        fileSize:
          type: ["integer", "null"]
        fileUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    DocumentInput:
      type: object
      required: [name, folder, fileType]
      properties:
        name:
          type: string
        folder:
          type: string
        fileType:
          type: string
        fileSize:
          type: ["integer", "null"]
        fileUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...