# lib/api-zod/src/generated/types/monthlyRevenue.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MEI Fácil API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyRevenue {
  month: number;
  year: number;
  total: number;
  label?: string;
}